swapped format to float instead of int

+++ supervised_learning/error_analysis/0-create_confusion.py
@@ -21,7 +21,7 @@ def create_confusion_matrix(labels, logits):
     true_classes = np.argmax(labels, axis=1)
     pred_classes = np.argmax(logits, axis=1)
     classes = labels.shape[1]
-    confusion = np.zeros((classes, classes), dtype=int)
+    confusion = np.zeros((classes, classes), dtype=float)
     for t, p in zip(true_classes, pred_classes):
         confusion[t, p] += 1
     return confusion
